fix: transaction-list swap and bridge unit tests
