+++ test/data/bridge/mock-bridge-store.ts
@@ -108,6 +108,7 @@ export const createBridgeMockStore = ({
     },
     localeMessages: { currentLocale: 'es_419' },
     metamask: {
+      featureFlags: {},
       ...DEFAULT_BRIDGE_STATUS_CONTROLLER_STATE,
       ...swapsMetamask,
       ...mockNetworkState(
